requirements.txt
tensorflow
pillow
flask
flask_cors
numpy
qdrant-client
cloudinary